+++ src/main/java/znu/visum/components/movies/usecases/markasfavorite/domain/MarkAsFavoriteMovieService.java
@@ -1,0 +1,5 @@
+package znu.visum.components.movies.usecases.markasfavorite.domain;
+
+public interface MarkAsFavoriteMovieService {
+  boolean markAsFavorite(long movieId);
+}