+++ src/main/java/znu/visum/components/movies/usecases/markastowatch/domain/MarkAsToWatchMovieService.java
@@ -1,0 +1,5 @@
+package znu.visum.components.movies.usecases.markastowatch.domain;
+
+public interface MarkAsToWatchMovieService {
+  boolean markAsToWatch(long movieId);
+}