+++ src/main/java/znu/visum/components/movies/usecases/removetowatch/domain/RemoveToWatchMovieService.java
@@ -1,0 +1,5 @@
+package znu.visum.components.movies.usecases.removetowatch.domain;
+
+public interface RemoveToWatchMovieService {
+  boolean removeToWatch(long movieId);
+}